split file for marshal and unmarshal
